pyproject.toml
[tool.poetry]
name = "rasierwasser"
version = "0.1.0"
description = "Simple pip repository server for internal usage"
authors = ["voidpointercast <patrick.daniel.gress@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
pydantic = "^1.8.2"
SQLAlchemy = "^1.4.17"
pyOpenSSL = "^20.0.1"
fastapi = "^0.65.1"
Jinja2 = "^3.0.1"

[tool.poetry.dev-dependencies]
uvicorn = "^0.14.0"
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
